+++ packages/styles/bases/search-input.css.ts
@@ -5,6 +5,9 @@ const borderWidth = "1px";
 
 export const searchStyle = style({
   position: "relative",
+  background: THEME.color.white["800"],
+  borderRadius: THEME.spacing.xs2,
+  boxSizing: "border-box",
 });
 
 const BORDER_WIDTH = "1px";
@@ -31,9 +34,6 @@ export const searchInputDisabledStyle = style({
 });
 
 export const searchInputIconStyle = style({
-  position: "absolute",
-  top: "50%",
-  transform: "translateY(-50%)",
   left: THEME.spacing.xs,
   display: "flex",
   alignItems: "center",
@@ -149,6 +149,10 @@ export const searchDropdownCheckboxItemStyle = style({
   padding: `${THEME.spacing.sm} ${THEME.spacing.xs2}`,
 });
 
+export const searchDropdownSingleSelectItemMarginStyle = style({
+  padding: `${THEME.spacing.sm} 0`,
+});
+
 export const searchCheckboxInputStyle = style({
   display: "none",
 });
@@ -200,4 +204,107 @@ export const searchInputCheckboxStyle = style({});
 // FIXME: WizCheckBoxNewで置換する際に消す。
 globalStyle(`${searchInputCheckboxStyle} > div`, {
   width: "100%",
+});
+
+export const searchDropdownSingleSelectItemStyle = style({
+  width: "100%",
+  display: "flex",
+  justifyContent: "space-between",
+  alignItems: "center",
+  cursor: "pointer",
+  fontSize: THEME.fontSize.sm,
+  userSelect: "none",
+  border: "none",
+  backgroundColor: THEME.color.transparent,
+  color: THEME.color.gray[800],
+  textAlign: "left",
+  padding: `0.375rem  ${THEME.spacing.xs}`,
+  "@media": {
+    "(any-hover: hover)": {
+      ":hover": {
+        background: THEME.color.green[300],
+        color: THEME.color.green[800],
+      },
+    },
+  },
+});
+
+export const searchInputInnerBoxStyle = style({
+  height: "100%",
+  padding: `calc(${THEME.spacing.xs} - ${BORDER_WIDTH}) ${THEME.spacing.xs}`,
+  fontSize: THEME.fontSize.sm,
+  color: THEME.color.gray["500"],
+  width: "100%",
+  boxSizing: "border-box",
+  display: "inline-block",
+});
+
+export const searchInputSelectedItemStyle = style({
+  height: "100%",
+  fontSize: THEME.fontSize.sm,
+  color: THEME.color.gray["500"],
+  // width: "100%",
+  boxSizing: "border-box",
+  display: "inline-block",
+});
+
+export const searchInputInnerBoxSelectedItemStyle = style({
+  overflow: "hidden",
+  textOverflow: "ellipsis",
+  whiteSpace: "nowrap",
+
+  display: "flex",
+  flexDirection: "row",
+  float: "left",
+  alignItems: "center",
+  padding: `${THEME.spacing.xs2} ${THEME.spacing.xs}`,
+  gap: THEME.spacing.xs,
+
+  background: THEME.color.gray["300"],
+  border: `1px solid ${THEME.color.gray["400"]}`,
+  borderRadius: THEME.spacing.xs2,
+
+  color: THEME.color.gray["800"],
+});
+
+export const searchInputInnerBoxSelectedLabelStyle = style({
+  lineHeight: THEME.fontSize.lg,
+  overflow: "hidden",
+  textOverflow: "ellipsis",
+  whiteSpace: "nowrap",
+  userSelect: "none",
+});
+
+export const searchInputInnerBoxCloseStyle = style({
+  fill: THEME.color.gray["500"],
+  fontSize: THEME.spacing.xs,
+});
+
+export const searchInputInnerBoxCloseButtonStyle = style({
+  border: "none",
+  background: "none",
+  padding: 0,
+  cursor: "pointer",
+  ":disabled": {
+    cursor: "not-allowed",
+  },
+});
+
+export const searchInputInnerInputStyle = style({
+  minWidth: "30%",
+  border: "none",
+  outline: "none",
+  padding: `${THEME.spacing.xs2} ${THEME.spacing.no}`,
+  lineHeight: THEME.fontSize.xl,
+  flexGrow: 1,
+  fontSize: THEME.fontSize.sm,
+  color: THEME.color.gray["800"],
+  "::placeholder": {
+    color: THEME.color.gray["500"],
+    userSelect: "none",
+  },
+  ":disabled": {
+    cursor: "not-allowed",
+    backgroundColor: THEME.color.gray["300"],
+  },
 });