+++ packages/wiz-ui-next/src/components/base/inputs/search-selector/search-selector.vue
@@ -28,6 +28,7 @@
               :class="selectBoxInnerBoxCloseButtonStyle"
               :ref="setUnselectableRef(i)"
               :aria-label="ARIA_LABELS.SEARCH_SELECTOR.UNSELECT"
+              :disabled="disabled"
             >
               <WizIcon
                 :icon="WizIClose"
@@ -88,7 +89,7 @@
         "
       >
         <WizVStack gap="xs2">
-          <WizPopupButtonGroup
+          <PopupButtonGroup
             :options="
               addButtonEnabled ? [addButton, ...selectButtons] : selectButtons
             "
@@ -119,12 +120,7 @@ import {
 import { inputBorderStyle } from "@wizleap-inc/wiz-ui-styles/commons";
 import { ComponentPublicInstance, PropType, computed, inject, ref } from "vue";
 
-import {
-  WizIcon,
-  WizPopup,
-  WizPopupButtonGroup,
-  WizPopupContainer,
-} from "@/components";
+import { WizIcon, WizPopup, WizPopupContainer } from "@/components";
 import {
   WizIAddCircle,
   WizIClose,
@@ -133,10 +129,10 @@ import {
 } from "@/components/icons";
 import { formControlKey } from "@/hooks/use-form-control-provider";
 
-import { ButtonGroupItem } from "../../popup-button-group/types";
 import { WizHStack, WizVStack } from "../../stack";
 
-import { levenshteinDistance } from "./levenshtein-distance";
+import { filterOptions } from "./levenshtein-distance";
+import { ButtonGroupItem, PopupButtonGroup } from "./popup-button-group";
 import { SelectBoxOption } from "./types";
 
 defineOptions({
@@ -211,6 +207,18 @@ const props = defineProps({
     type: String,
     required: false,
   },
+  /**
+   * 検索対象に含む類似度の閾値を，0から1の範囲で指定します。
+   * 類似度は標準化レーベンシュタイン距離に基づいて計算され，0に近いほど類似しています。
+   * ただし，類似度の最小値が閾値を上回る場合は部分一致で検索します。
+   * @default 0.75
+   *
+   */
+  threshold: {
+    type: Number,
+    required: false,
+    default: 0.75,
+  },
 });
 
 const emit = defineEmits<Emit>();
@@ -228,26 +236,6 @@ const toggleDropdown = () => {
 };
 
 const deepCopy = <T>(ary: T): T => JSON.parse(JSON.stringify(ary));
-
-const selectByLevenshteinAndPartialMatch = (
-  options: SelectBoxOption[],
-  target: string
-) => {
-  const dist = options.reduce((acc, str) => {
-    acc[str.label] = levenshteinDistance(str.label, target);
-    return acc;
-  }, {} as { [key: string]: number });
-  const minLength = Math.min(...Object.values(dist));
-  const closestWords = options.filter(
-    (option) => dist[option.label] === minLength
-  );
-
-  const exactMatch = options.filter((option) => {
-    const isIncluded = option.label.indexOf(target) !== -1;
-    return isIncluded && !closestWords.includes(option);
-  });
-  return closestWords.concat(exactMatch);
-};
 
 const valueToOption = computed(() =>
   props.options.reduce((acc, item) => {
@@ -272,12 +260,16 @@ const setUnselectableRef =
 
 const filteredOptions = computed(() => {
   const sortedOptions =
-    props.searchValue.length !== 0
-      ? selectByLevenshteinAndPartialMatch(
+    props.searchValue.length === 0
+      ? props.options
+      : filterOptions(
           deepCopy(props.options),
-          props.searchValue
-        )
-      : props.options;
+          props.searchValue,
+          props.threshold
+        ).filter(
+          (matchedOption) =>
+            !props.modelValue.some((value) => matchedOption.value === value)
+        );
   const removeSelectedOptions = (options: SelectBoxOption[]) => {
     return options.filter((v) => {
       return !selectedItem.value.some((item) => item.value === v.value);
@@ -387,6 +379,7 @@ const selectButtons = computed(() => {
         label: opt.label,
         value: opt.value,
         exLabel: opt.exLabel,
+        disabled: opt.disabled,
         onClick: () => onSelect(opt.value),
       },
     };