+++ packages/wiz-ui-next/src/components/base/snackbar/snackbar-controller.vue
@@ -25,13 +25,14 @@ const { snackbarOptions } = globalInject(globalKey);
 const containerRef = ref<HTMLElement | undefined>();
 
 const space2px = (key: SpacingKeys) => {
+  if (!window) return 0;
   const rem = getSpacingCss(key);
   if (!rem || rem === "0") return 0;
   if (rem === "auto") return 0;
   if (rem === "9999px") return 9999;
   return (
     parseFloat(rem) *
-    parseFloat(getComputedStyle(document.documentElement).fontSize)
+    parseFloat(window.getComputedStyle(document.documentElement).fontSize)
   );
 };