+++ packages/wiz-ui-next/src/components/icons/index.ts
@@ -1,9 +1,13 @@
 import { default as WizIAddCircle } from "./add-circle.vue";
 import { default as WizIAdd } from "./add.vue";
+import { default as WizIAdjustBold } from "./adjust-bold.vue";
+import { default as WizIAdjust } from "./adjust.vue";
 import { default as WizIApprovalFilled } from "./approval-filled.vue";
+import { default as WizIArrowDownward } from "./arrow-downward.vue";
 import { default as WizIArrowDropDown } from "./arrow-drop-down.vue";
 import { default as WizIArrowDropUp } from "./arrow-drop-up.vue";
 import { default as WizIArrowRight } from "./arrow-right.vue";
+import { default as WizIArrowUpward } from "./arrow-upward.vue";
 import { default as WizIAssignmentAdd } from "./assignment-add.vue";
 import { default as WizIAssignment } from "./assignment.vue";
 import { default as WizIAttachFile } from "./attach-file.vue";
@@ -33,6 +37,7 @@ import { default as WizIContactPageFilled } from "./contact-page-filled.vue";
 import { default as WizIContentCopy } from "./content-copy.vue";
 import { default as WizIContract } from "./contract.vue";
 import { default as WizIDashboard } from "./dashboard.vue";
+import { default as WizIDelete } from "./delete.vue";
 import { default as WizIDemography } from "./demography.vue";
 import { default as WizIDescription } from "./description.vue";
 import { default as WizIDomain } from "./domain.vue";
@@ -40,23 +45,30 @@ import { default as WizIDownload } from "./download.vue";
 import { default as WizIDraft } from "./draft.vue";
 import { default as WizIDrafts } from "./drafts.vue";
 import { default as WizIEditNote } from "./edit-note.vue";
+import { default as WizIEmergencyHome } from "./emergency-home.vue";
 import { default as WizIExpandLess } from "./expand-less.vue";
 import { default as WizIExpandMoreBold } from "./expand-more-bold.vue";
 import { default as WizIExpandMore } from "./expand-more.vue";
 import { default as WizIEye } from "./eye.vue";
 import { default as WizIFamilyHome } from "./family-home.vue";
+import { default as WizIFeedback } from "./feedback.vue";
 import { default as WizIForwardToInbox } from "./forward-to-inbox.vue";
 import { default as WizIGroups } from "./groups.vue";
 import { default as WizIHelp } from "./help.vue";
 import { default as WizIHistory } from "./history.vue";
 import { default as WizIHome } from "./home.vue";
+import { default as WizIHourglassTop } from "./hourglass-top.vue";
+import { default as WizIInboxCustomize } from "./inbox-customize.vue";
 import { default as WizIInfoBold } from "./info-bold.vue";
 import { default as WizIInfoFilled } from "./info-filled.vue";
 import { default as WizIInfo } from "./info.vue";
+import { default as WizILinkOff } from "./link-off.vue";
 import { default as WizILocationCity } from "./location-city.vue";
 import { default as WizILocationOn } from "./location-on.vue";
 import { default as WizILogin } from "./login.vue";
+import { default as WizIMail } from "./mail.vue";
 import { default as WizIManageAccounts } from "./manage-accounts.vue";
+import { default as WizIMenuOpen } from "./menu-open.vue";
 import { default as WizIMenu } from "./menu.vue";
 import { default as WizIMonitoring } from "./monitoring.vue";
 import { default as WizIMoreVert } from "./more-vert.vue";
@@ -69,6 +81,7 @@ import { default as WizIPinDrop } from "./pin-drop.vue";
 import { default as WizIPublic } from "./public.vue";
 import { default as WizIRemoveBold } from "./remove-bold.vue";
 import { default as WizIRemove } from "./remove.vue";
+import { default as WizIReviews } from "./reviews.vue";
 import { default as WizISchedule } from "./schedule.vue";
 import { default as WizISearch } from "./search.vue";
 import { default as WizISend } from "./send.vue";
@@ -87,10 +100,14 @@ import { default as WizIWarning } from "./warning.vue";
 export type TIcon =
   | typeof WizIAddCircle
   | typeof WizIAdd
+  | typeof WizIAdjustBold
+  | typeof WizIAdjust
   | typeof WizIApprovalFilled
+  | typeof WizIArrowDownward
   | typeof WizIArrowDropDown
   | typeof WizIArrowDropUp
   | typeof WizIArrowRight
+  | typeof WizIArrowUpward
   | typeof WizIAssignmentAdd
   | typeof WizIAssignment
   | typeof WizIAttachFile
@@ -120,6 +137,7 @@ export type TIcon =
   | typeof WizIContentCopy
   | typeof WizIContract
   | typeof WizIDashboard
+  | typeof WizIDelete
   | typeof WizIDemography
   | typeof WizIDescription
   | typeof WizIDomain
@@ -127,23 +145,30 @@ export type TIcon =
   | typeof WizIDraft
   | typeof WizIDrafts
   | typeof WizIEditNote
+  | typeof WizIEmergencyHome
   | typeof WizIExpandLess
   | typeof WizIExpandMoreBold
   | typeof WizIExpandMore
   | typeof WizIEye
   | typeof WizIFamilyHome
+  | typeof WizIFeedback
   | typeof WizIForwardToInbox
   | typeof WizIGroups
   | typeof WizIHelp
   | typeof WizIHistory
   | typeof WizIHome
+  | typeof WizIHourglassTop
+  | typeof WizIInboxCustomize
   | typeof WizIInfoBold
   | typeof WizIInfoFilled
   | typeof WizIInfo
+  | typeof WizILinkOff
   | typeof WizILocationCity
   | typeof WizILocationOn
   | typeof WizILogin
+  | typeof WizIMail
   | typeof WizIManageAccounts
+  | typeof WizIMenuOpen
   | typeof WizIMenu
   | typeof WizIMonitoring
   | typeof WizIMoreVert
@@ -156,6 +181,7 @@ export type TIcon =
   | typeof WizIPublic
   | typeof WizIRemoveBold
   | typeof WizIRemove
+  | typeof WizIReviews
   | typeof WizISchedule
   | typeof WizISearch
   | typeof WizISend
@@ -174,10 +200,14 @@ export type TIcon =
 export {
   WizIAddCircle,
   WizIAdd,
+  WizIAdjustBold,
+  WizIAdjust,
   WizIApprovalFilled,
+  WizIArrowDownward,
   WizIArrowDropDown,
   WizIArrowDropUp,
   WizIArrowRight,
+  WizIArrowUpward,
   WizIAssignmentAdd,
   WizIAssignment,
   WizIAttachFile,
@@ -207,6 +237,7 @@ export {
   WizIContentCopy,
   WizIContract,
   WizIDashboard,
+  WizIDelete,
   WizIDemography,
   WizIDescription,
   WizIDomain,
@@ -214,23 +245,30 @@ export {
   WizIDraft,
   WizIDrafts,
   WizIEditNote,
+  WizIEmergencyHome,
   WizIExpandLess,
   WizIExpandMoreBold,
   WizIExpandMore,
   WizIEye,
   WizIFamilyHome,
+  WizIFeedback,
   WizIForwardToInbox,
   WizIGroups,
   WizIHelp,
   WizIHistory,
   WizIHome,
+  WizIHourglassTop,
+  WizIInboxCustomize,
   WizIInfoBold,
   WizIInfoFilled,
   WizIInfo,
+  WizILinkOff,
   WizILocationCity,
   WizILocationOn,
   WizILogin,
+  WizIMail,
   WizIManageAccounts,
+  WizIMenuOpen,
   WizIMenu,
   WizIMonitoring,
   WizIMoreVert,
@@ -243,6 +281,7 @@ export {
   WizIPublic,
   WizIRemoveBold,
   WizIRemove,
+  WizIReviews,
   WizISchedule,
   WizISearch,
   WizISend,