+++ packages/wiz-ui-react/src/components/base/inputs/search-input/components/search-popup-panel.tsx
@@ -19,7 +19,9 @@ type Props = BaseProps & {
   values: number[];
   width?: string;
   emptyMessage: string;
+  singleSelect?: boolean;
   onChangeValues: (values: number[]) => void;
+  closePopup: () => void;
 };
 
 export const SearchPopupPanel: FC<Props> = ({
@@ -29,7 +31,9 @@ export const SearchPopupPanel: FC<Props> = ({
   values,
   width,
   emptyMessage,
+  singleSelect,
   onChangeValues,
+  closePopup,
 }) => {
   const [activeValue, setActiveValue] = useState<number | null>(null);
   const activeOption = useMemo(
@@ -54,10 +58,17 @@ export const SearchPopupPanel: FC<Props> = ({
     [onChangeValues, values]
   );
 
+  const handleClickButton = (value: number) => {
+    onChangeValues([value]);
+    closePopup();
+  };
+
   useEffect(() => {
     // reset active
     setActiveValue(null);
   }, [options]);
+
+  const allOptionsHaveChildren = options.every((option) => !!option.children);
 
   return (
     <>
@@ -94,6 +105,9 @@ export const SearchPopupPanel: FC<Props> = ({
                       )}
                     >
                       <WizHStack
+                        pl={
+                          !allOptionsHaveChildren && !singleSelect ? "lg" : "no"
+                        }
                         py="xs2"
                         width="100%"
                         justify="between"
@@ -123,32 +137,64 @@ export const SearchPopupPanel: FC<Props> = ({
                     </div>
                   </div>
                 ) : (
-                  // CheckBox
-                  <div className={styles.searchDropdownCheckboxItemStyle}>
-                    <WizCheckBoxNew
-                      style={{ width: "100%" }}
-                      value={option.value}
-                      id={`${option.label}-${option.value}`}
-                      checked={values.includes(option.value)}
-                      onChange={(e) => {
-                        handleChangeValues(option.value, e.target.checked);
-                      }}
-                    >
-                      <WizHStack width="100%" align="center" gap="xs2" nowrap>
-                        <div className={styles.searchInputLabelStyle}>
-                          {option.label}
-                        </div>
-                        {optionTag && (
-                          <WizTag
-                            label={optionTag.label}
-                            variant="white"
-                            width="20px"
-                            fontSize="xs2"
-                          />
-                        )}
-                      </WizHStack>
-                    </WizCheckBoxNew>
-                  </div>
+                  <>
+                    {singleSelect ? (
+                      <div className={styles.searchDropdownItemStyle}>
+                        <button
+                          id={`${option.label}-${option.value}`}
+                          style={{ width: "100%" }}
+                          type="button"
+                          className={styles.searchDropdownSingleSelectItemStyle}
+                          onClick={() => handleClickButton(option.value)}
+                        >
+                          <div className={styles.searchInputLabelStyle}>
+                            {option.label}
+                          </div>
+                          {optionTag ? (
+                            <WizTag
+                              label={optionTag.label}
+                              variant="white"
+                              width="20px"
+                              fontSize="xs2"
+                            />
+                          ) : (
+                            <div style={{ width: "1.5rem" }} />
+                          )}
+                        </button>
+                      </div>
+                    ) : (
+                      <div className={styles.searchDropdownCheckboxItemStyle}>
+                        <WizCheckBoxNew
+                          style={{ width: "100%" }}
+                          value={option.value}
+                          id={`${option.label}-${option.value}`}
+                          checked={values.includes(option.value)}
+                          onChange={(e) => {
+                            handleChangeValues(option.value, e.target.checked);
+                          }}
+                        >
+                          <WizHStack
+                            width="100%"
+                            align="center"
+                            gap="xs2"
+                            nowrap
+                          >
+                            <div className={styles.searchInputLabelStyle}>
+                              {option.label}
+                            </div>
+                            {optionTag && (
+                              <WizTag
+                                label={optionTag.label}
+                                variant="white"
+                                width="20px"
+                                fontSize="xs2"
+                              />
+                            )}
+                          </WizHStack>
+                        </WizCheckBoxNew>
+                      </div>
+                    )}
+                  </>
                 )}
                 <WizDivider color="gray.300" />
               </div>
@@ -165,6 +211,8 @@ export const SearchPopupPanel: FC<Props> = ({
           options={activeOptionChildren}
           values={values}
           width={width}
+          closePopup={closePopup}
+          singleSelect={singleSelect}
           emptyMessage={emptyMessage}
           onChangeValues={onChangeValues}
         />