+++ packages/wiz-ui-react/src/components/custom/chat/stories/chat-card.stories.tsx
@@ -86,6 +86,7 @@ const dummyStatusOptions: ComponentProps<typeof WizChatCard>["statusOptions"] =
   ];
 
 const templateArgs = {
+  isOpen: true,
   username: "マネーキャリアスタッフ",
   placeholder: "入力してください",
   messages: dummyMessages,
@@ -140,6 +141,7 @@ export const Status: Story = {
 export const Closed: Story = {
   args: {
     ...templateArgs,
+    isOpen: false,
   },
 };
 
@@ -147,6 +149,7 @@ export const HaveNewMessage: Story = {
   args: {
     ...templateArgs,
     haveNewMessage: true,
+    isOpen: false,
   },
 };
 
@@ -156,9 +159,11 @@ export const Playground: Story = {
     const [textValue, setTextValue] = useState("");
     const [haveNewMessage, setHaveNewMessage] = useState(true);
     const [status, setStatus] = useState<number | null>(null);
+    const [isOpen, setIsOpen] = useState(false);
     return (
       <WizChatCard
         messages={messages}
+        isOpen={isOpen}
         haveNewMessage={haveNewMessage}
         textValue={textValue}
         username="マネーキャリアスタッフ"
@@ -167,7 +172,10 @@ export const Playground: Story = {
         typingUsername="なんとかかんとか"
         status={status}
         statusOptions={dummyStatusOptions}
-        onToggleOpen={() => setHaveNewMessage(false)}
+        onToggle={() => {
+          setHaveNewMessage(false);
+          setIsOpen((current) => !current);
+        }}
         onChangeTextValue={(changed) => setTextValue(changed)}
         onChangeStatus={(changed) => setStatus(changed)}
         onSubmit={() => {